fixed bug with scale not written to DB - Improved Point-Selection to support double-values

+++ W-Seminar/src/pointSelection/PictureLine.java
@@ -5,15 +5,15 @@ import java.awt.Color;
 import javax.swing.JPanel;
 
 public class PictureLine {
-	private int x1; 
-    private int y1;
-    private int x2;
-    private int y2;   
+	private double x1; 
+    private double y1;
+    private double x2;
+    private double y2;   
     private Color color;
     private JPanel view;
     
     
-    public PictureLine(int x1, int y1, int x2, int y2, Color color, JPanel parent) {
+    public PictureLine(double x1, double y1, double x2, double y2, Color color, JPanel parent) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
@@ -26,7 +26,7 @@ public class PictureLine {
 	/**
 	 * @return the x1
 	 */
-	public int getX1() {
+	public double getX1() {
 		return x1;
 	}
 
@@ -34,7 +34,7 @@ public class PictureLine {
 	/**
 	 * @param x1 the x1 to set
 	 */
-	public void setX1(int x1) {
+	public void setX1(double x1) {
 		this.x1 = x1;
 		view.repaint();
 	}
@@ -43,7 +43,7 @@ public class PictureLine {
 	/**
 	 * @return the y1
 	 */
-	public int getY1() {
+	public double getY1() {
 		return y1;
 	}
 
@@ -51,7 +51,7 @@ public class PictureLine {
 	/**
 	 * @param y1 the y1 to set
 	 */
-	public void setY1(int y1) {
+	public void setY1(double y1) {
 		this.y1 = y1;
 		view.repaint();
 	}
@@ -60,7 +60,7 @@ public class PictureLine {
 	/**
 	 * @return the x2
 	 */
-	public int getX2() {
+	public double getX2() {
 		return x2;
 	}
 
@@ -68,7 +68,7 @@ public class PictureLine {
 	/**
 	 * @param x2 the x2 to set
 	 */
-	public void setX2(int x2) {
+	public void setX2(double x2) {
 		this.x2 = x2;
 		view.repaint();
 	}
@@ -77,7 +77,7 @@ public class PictureLine {
 	/**
 	 * @return the y2
 	 */
-	public int getY2() {
+	public double getY2() {
 		return y2;
 	}
 
@@ -85,7 +85,7 @@ public class PictureLine {
 	/**
 	 * @param y2 the y2 to set
 	 */
-	public void setY2(int y2) {
+	public void setY2(double y2) {
 		this.y2 = y2;
 		view.repaint();
 	}
